openapi_type

Trait Visitor

pub trait Visitor: Sealed {
    type OptionVisitor: Visitor;
    type ArrayVisitor: Visitor;
    type ObjectVisitor: ObjectVisitor;
    type AlternativesVisitor: AlternativesVisitor;

    // Required methods
    fn visit_unit_struct(
        &mut self,
        name: Option<String>,
        description: Option<String>,
    );
    fn visit_any(&mut self);
    fn visit_bool(&mut self);
    fn visit_int(&mut self, byte: Option<u32>, minimum: Option<i64>);
    fn visit_number(&mut self, byte: Option<u32>);
    fn visit_char(&mut self);
    fn visit_string(&mut self);
    fn visit_uuid(&mut self);
    fn visit_date(&mut self);
    fn visit_datetime(&mut self);
    fn visit_binary(&mut self);
    fn visit_option(&mut self) -> &mut Self::OptionVisitor;
    fn visit_enum<I>(
        &mut self,
        name: Option<String>,
        description: Option<String>,
        variants: I,
    )
       where I: IntoIterator<Item = String>;
    fn visit_array(
        &mut self,
        len: Option<usize>,
        unique_items: bool,
    ) -> &mut Self::ArrayVisitor;
    fn visit_object(&mut self) -> &mut Self::ObjectVisitor;
    fn visit_alternatives(&mut self) -> &mut Self::AlternativesVisitor;

    // Provided method
    fn visit_unit(&mut self) { ... }

}

This trait can be used to visit a type. Call one of the methods on this trait exactly once.

Required Associated Types

type OptionVisitor: Visitor

type ArrayVisitor: Visitor

type ObjectVisitor: ObjectVisitor

type AlternativesVisitor: AlternativesVisitor

Required Methods

fn visit_unit_struct( &mut self, name: Option<String>, description: Option<String>, )

fn visit_any(&mut self)

fn visit_bool(&mut self)

fn visit_int(&mut self, byte: Option<u32>, minimum: Option<i64>)

fn visit_number(&mut self, byte: Option<u32>)

fn visit_char(&mut self)

fn visit_string(&mut self)

fn visit_uuid(&mut self)

fn visit_date(&mut self)

fn visit_datetime(&mut self)

fn visit_binary(&mut self)

fn visit_option(&mut self) -> &mut Self::OptionVisitor

fn visit_enum<I>( &mut self, name: Option<String>, description: Option<String>, variants: I, )
where I: IntoIterator<Item = String>,

fn visit_array( &mut self, len: Option<usize>, unique_items: bool, ) -> &mut Self::ArrayVisitor

fn visit_object(&mut self) -> &mut Self::ObjectVisitor

fn visit_alternatives(&mut self) -> &mut Self::AlternativesVisitor

Provided Methods

fn visit_unit(&mut self)

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.